Fixes a couple of issues
- Creating a new mnemonic without saving it
- Adding the ability to hide the "button" when not needed. This prevents
the button from staying around because it's also transparent and then
clicking it from within the app is possible leading to a "sign-in"
triggered after this.

The button was made that way so that it works from within the iframe in
iOS. In which the popup window/tab was not being created.