Fixes some issues
This fixes the initial bug that prevented signing in from Google Chrome.

It also addresses a very strange bug that happens on Firefox when the
user has enhanced tracking protection.

Currently the solution we were able to implement is very good. We detect
that the sign in failed by intercepting one of the messages that the
gapi library exchanges with the main window (in our case the custodian
frame) and if we find the error and that it has occurred at the wrong
time then we assume that there was a problem signing in.

This does not imply that the failure was caused by said reason. But it's
certainly a string indicator. So instead of bothering the user with a
big message and such, we just show them a suggestion to check their
privacy settings.